Fix build error.
Add DosPadUIApplication files to project.

+++ trunk/SDL/include/SDL_opengles.h
@@ -105,8 +105,8 @@ extern "C"
     typedef int GLfixed;
     typedef int GLclampx;
 
-    typedef int GLintptr;
-    typedef int GLsizeiptr;
+    typedef long GLintptr;
+    typedef long GLsizeiptr;
 
 
 /*************************************************************/
